Rename CV Tools file for better readability #125
