+++ ansible_collections/arista/cvp/plugins/modules/cv_device.py
@@ -31,8 +31,8 @@ ANSIBLE_METADATA = {
 import logging
 import ansible_collections.arista.cvp.plugins.module_utils.logger   # noqa # pylint: disable=unused-import
 from ansible.module_utils.basic import AnsibleModule
-from ansible_collections.arista.cvp.plugins.module_utils.cv_tools import cv_update_configlets_on_device
-from ansible_collections.arista.cvp.plugins.module_utils.cv_tools import cv_connect, HAS_CVPRAC
+from ansible_collections.arista.cvp.plugins.module_utils.tools_cv import cv_update_configlets_on_device
+from ansible_collections.arista.cvp.plugins.module_utils.tools_cv import cv_connect, HAS_CVPRAC
 
 
 DOCUMENTATION = r"""
