+++ ansible_collections/arista/cvp/plugins/modules/cv_facts.py
@@ -30,14 +30,11 @@ import logging
 import traceback  # noqa # pylint: disable=unused-import
 import ansible_collections.arista.cvp.plugins.module_utils.logger   # noqa # pylint: disable=unused-import
 from ansible.module_utils.basic import AnsibleModule
-# from ansible_collections.arista.cvp.plugins.module_utils.cv_client import CvpClient
-# from ansible_collections.arista.cvp.plugins.module_utils.cv_client_errors import CvpLoginError
-
 from ansible_collections.arista.cvp.plugins.module_utils.tools_inventory import (
     find_hostname_by_mac,
     find_containerName_by_containerId
 )
-from ansible_collections.arista.cvp.plugins.module_utils.cv_tools import cv_connect, HAS_CVPRAC
+from ansible_collections.arista.cvp.plugins.module_utils.tools_cv import cv_connect, HAS_CVPRAC
 
 DOCUMENTATION = r'''
 ---
